projects/03_django_rest_framework/09_viewsets_and_routers/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
dj-rest-auth==7.0.0
Django==5.1.1
django-allauth==65.2.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.10
packaging==24.1
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.7.0
